+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 * Fixed problem with hidden publish and save buttons on iOS
 * Fixed problem with freezing page tree when moving pages in admin
 * Fixed problem with editor for file properties inside collections
+* Added option to specify connection string name in configuration
 
 ### New in 1.2.3
 * Fixed script reference problem for property editors in Chrome